requirements.txt
boto3==1.34.67
botocore==1.34.67
fire==0.6.0
psycopg2_binary==2.9.9
Requests==2.31.0
